refactor: move wiki component to new file
