+++ src/routes/(app)/space/[space]/[channel]/+page.svelte
@@ -31,11 +31,7 @@
   import { getProfile } from "$lib/profile.svelte";
 
   import { onMount, onDestroy } from "svelte";
-  import {
-    BlockNoteEditor,
-    type BlockNoteEditorOptions,
-  } from "@blocknote/core";
-  import "@blocknote/core/style.css";
+  import WikiEditor from "$lib/components/WikiEditor.svelte";
 
   let isMobile = $derived((outerWidth.current ?? 0) < 640);
 
@@ -223,30 +219,6 @@
   async function sendMessage() {
     if (!space) return;
 
-    /* TODO: rework with tiptap?
-    const images = imageFiles
-      ? await Promise.all(
-          Array.from(imageFiles).map(async (file) => {
-            try {
-              const resp = await user.uploadBlob(file);
-              return {
-                source: resp.url, // Use the blob reference from ATP
-                alt: file.name,
-              };
-            } catch (error) {
-              console.error("Failed to upload image:", error);
-              toast.error("Failed to upload image", { position: "bottom-end" });
-              return null;
-            }
-          }),
-        ).then((results) =>
-          results.filter(
-            (result): result is NonNullable<typeof result> => result !== null,
-          ),
-        )
-      : undefined;
-    */
-
     space.change((doc) => {
       if (!user.agent) return;
 
@@ -256,9 +228,6 @@
         reactions: {},
         content: JSON.stringify(messageInput),
         ...(replyingTo && { replyTo: replyingTo.id }),
-
-        // TODO: rework images with tiptap
-        // ...(images && { images }),
       };
       doc.channels[page.params.channel].timeline.push(id);
     });
@@ -345,378 +314,6 @@
     });
     showSettingsDialog = false;
   }
-
-  // TODO: rework with tiptap
-  async function handleImageSelect(event: Event) {
-    const input = event.target as HTMLInputElement;
-    imageFiles = input.files;
-  }
-
-  async function handlePaste(event: ClipboardEvent) {
-    if (!mayUploadImages) return;
-
-    const items = event.clipboardData?.items;
-    if (!items) return;
-
-    for (const item of items) {
-      if (item.type.startsWith("image/")) {
-        const file = item.getAsFile();
-        if (file) {
-          imageFiles = new DataTransfer().files;
-          const dataTransfer = new DataTransfer();
-          dataTransfer.items.add(file);
-          imageFiles = dataTransfer.files;
-        }
-      }
-    }
-  }
-
-  let wikiContent = $state(channel?.wiki?.content || null);
-  let wikiRenderedHtml = $state(channel?.wiki?.html || "");
-  let isEditingWiki = $state(false);
-  let editorElement: HTMLElement;
-  let editor: BlockNoteEditor | null;
-
-  let wikiRelatedThreads = $derived(() => {
-    if (!space || !channel?.wiki?.relatedThreads) return {};
-
-    const threads = {};
-    for (const threadId of channel.wiki.relatedThreads) {
-      if (space.view.threads[threadId]) {
-        threads[threadId] = space.view.threads[threadId];
-      }
-    }
-    return threads;
-  });
-
-  function setEditingWiki(value: boolean) {
-    // Only allow admins to edit the wiki
-    if (value && !isAdmin) {
-      toast.error("Only admins can edit the wiki", { position: "bottom-end" });
-      return;
-    }
-
-    isEditingWiki = value;
-
-    if (value) {
-      if (editor) {
-        editor = null;
-      }
-
-      // Use a longer timeout to ensure DOM is ready
-      setTimeout(initBlockNoteEditor, 100);
-    } else if (editor) {
-      editor = null;
-    }
-  }
-
-  const EditorHandler = (editor: BlockNoteEditor) => {
-    try {
-      const cursorPosition = editor.getTextCursorPosition();
-      if (!cursorPosition) return;
-
-      const block = cursorPosition.block;
-      const content = block.content?.[0];
-
-      if (content && content.type === "text" && content.text === "/") {
-        // Show the slash command menu
-        slashMenuVisible = true;
-
-        const { top, left } = editor.getSelectionBoundingBox()?.toJSON();
-
-        // TODO: Calculate the menu height and position it correctly
-        const menuHeight = 300; // Approximate max height of the menu
-        const viewportHeight = window.innerHeight;
-        const scrollY = window.scrollY;
-
-        const wouldOverflow = top + 20 + menuHeight > viewportHeight + scrollY;
-
-        // Position menu above cursor if it would overflow, otherwise below
-        const yPosition = wouldOverflow
-          ? Math.max(scrollY, top - menuHeight - 10) // Above cursor, but not above viewport
-          : top + 20; // Below cursor with padding
-
-        slashMenuPosition = {
-          x: left,
-          y: yPosition,
-        };
-      } else {
-        slashMenuVisible = false;
-      }
-    } catch (e) {
-      console.error("Error in EditorHandler:", e);
-      slashMenuVisible = false;
-    }
-  };
-
-  const SelectionHandler = (editor: BlockNoteEditor) => {
-    try {
-      const selection = editor.getSelection();
-      
-      if (selection) {
-        const rect = editor.getSelectionBoundingBox()?.toJSON();
-        if (rect) {
-          selectionTooltipPosition = {
-            x: rect.left + (rect.width / 2) , 
-            y: rect.top - 50 
-          };
-          
-          selectionTooltipVisible = true;
-        }
-      } else {
-        selectionTooltipVisible = false;
-      }
-    } catch (e) {
-      console.error("Error in SelectionHandler:", e);
-      selectionTooltipVisible = false;
-    }
-  };
-
-  // Function to execute a slash command
-  function executeSlashCommand(command) {
-    if (!editor) return;
-
-    // Remove the slash character
-    const pos = editor.getTextCursorPosition();
-    
-
-    // Execute the command
-    command.action();
-    editor.updateBlock(pos.block.id, {
-      content: [],
-    });
-    editor.setTextCursorPosition(pos.block.id, "start");
-    // Hide the menu
-    slashMenuVisible = false;
-  }
-
-  async function initBlockNoteEditor() {
-    if (!editorElement) return;
-
-    try {
-      editorElement.innerHTML = "";
-
-      // Create the editor
-      editor = BlockNoteEditor.create();
-      editor.mount(editorElement);
-      editor.onChange(EditorHandler);
-      editor.onSelectionChange(SelectionHandler);
-      let targetBlockId = null;
-      
-      // Add click event listener for the "+" button pseudo-elements
-      editorElement.addEventListener('click', (e) => {
-        // Calculate if click was in the "+" button area
-        const blockElement = e.target.closest('.bn-block');
-        if (blockElement) {
-          const blockRect = blockElement.getBoundingClientRect();
-          
-          if (e.clientX < blockRect.left + 24) {
-            e.preventDefault();
-            e.stopPropagation();
-            
-            const blockId = blockElement.getAttribute('data-id');
-            if (!blockId) return;
-            
-            slashMenuPosition = {
-              x: blockRect.left + 30, 
-              y: e.clientY
-            };
-            
-            slashMenuVisible = true;
-            
-            targetBlockId = blockId;
-          }
-        }
-      });
-      
-
-      // Load existing content if available
-      if (wikiContent) {
-        try {
-          // Parse the content and set it to the editor
-          const parsedContent = JSON.parse(wikiContent);
-          console.log("Parsed content:", parsedContent);
-
-          // Wait a bit for the editor to fully initialize
-          setTimeout(() => {
-            if (editor && editor.document) {
-              // Replace the editor's document with the saved content
-              editor.replaceBlocks(editor.document, parsedContent);
-              console.log("Content loaded into editor");
-            }
-          }, 200);
-        } catch (e) {
-          console.error("Failed to parse wiki content", e);
-        }
-      }
-    } catch (e) {
-      console.error("Failed to initialize BlockNote editor", e);
-    }
-  }
-
-  async function saveWikiContent() {
-    if (!editor || !space) return;
-
-    try {
-      const content = JSON.stringify(editor.document);
-
-      const html = await editor.blocksToFullHTML(editor.document);
-      console.log(html);
-
-      space.change((doc) => {
-        if (!doc.channels[page.params.channel].wiki) {
-          doc.channels[page.params.channel].wiki = {
-            content,
-            html,
-            relatedThreads: [],
-            lastUpdated: Date.now(),
-          };
-        } else {
-          if (!doc.channels[page.params.channel].wiki) {
-            doc.channels[page.params.channel].wiki = {
-              content: "",
-              html: "",
-              relatedThreads: [],
-              lastUpdated: Date.now(),
-            };
-          }
-          doc.channels[page.params.channel].wiki.content = content;
-          doc.channels[page.params.channel].wiki.html = html;
-          doc.channels[page.params.channel].wiki.lastUpdated = Date.now();
-        }
-      });
-
-      wikiContent = content;
-      wikiRenderedHtml = html;
-      setEditingWiki(false);
-      toast.success("Wiki saved successfully", { position: "bottom-end" });
-    } catch (e) {
-      console.error("Failed to save wiki content", e);
-      toast.error("Failed to save wiki content", { position: "bottom-end" });
-    }
-  }
-
-  $effect(() => {
-    if (channel?.wiki) {
-      wikiContent = channel.wiki.content;
-      wikiRenderedHtml = channel.wiki.html || "";
-    }
-  });
-
-  // Clean up the editor when component unmounts or tab changes
-  $effect(() => {
-    if (tab !== "wiki" && editor) {
-      editor = null;
-    }
-  });
-
-  onMount(() => {
-    // Force reload CSS to ensure styles are applied
-    const existingLink = document.querySelector(
-      'link[href*="@blocknote/core/style.css"]',
-    );
-    if (!existingLink) {
-      const link = document.createElement("link");
-      link.rel = "stylesheet";
-      link.href = "https://unpkg.com/@blocknote/core@latest/style.css";
-      document.head.appendChild(link);
-    }
-  });
-
-  // Add these state variables for slash commands
-  let slashMenuVisible = $state(false);
-  let slashMenuPosition = $state({ x: 0, y: 0 });
-  let slashCommands = $state([
-    {
-      name: "Heading 1",
-      icon: "tabler:h-1",
-      action: () => editor?.updateBlock(editor?.getTextCursorPosition()?.block.id, { type: "heading", props: {level: 1} }),
-    },
-    {
-      name: "Heading 2",
-      icon: "tabler:h-2",
-      action: () => editor?.updateBlock(editor?.getTextCursorPosition()?.block.id, { type: "heading", props: {level: 2} }),
-    },
-    {
-      name: "Heading 3",
-      icon: "tabler:h-3",
-      action: () => editor?.updateBlock(editor?.getTextCursorPosition()?.block.id, { type: "heading", props: {level: 3} }),
-    },
-    {
-      name: "Bulleted List",
-      icon: "tabler:list",
-      action: () => editor?.updateBlock(editor?.getTextCursorPosition()?.block.id, { type: "bulletListItem" }),
-    },
-    {
-      name: "Numbered List",
-      icon: "tabler:list-numbers",
-      action: () => editor?.updateBlock(editor?.getTextCursorPosition()?.block.id, { type: "numberedListItem" }),
-    },
-    {
-      name: "Checklist",
-      icon: "tabler:checkbox",
-      action: () => editor?.updateBlock(editor?.getTextCursorPosition()?.block.id, { type: "checkListItem" }),
-    },
-    {
-      name: "Code Block",
-      icon: "tabler:code",
-      action: () => editor?.updateBlock(editor?.getTextCursorPosition()?.block.id, { type: "codeBlock" }),
-    }
-  ]);
-  function addLinkToSelection() {
-    if (!editor) return;
-    
-    const selection = editor.getSelection();
-    if (!selection) {
-      toast.error("Please select text to create a link", { position: "bottom-end" });
-      return;
-    }
-    
-    urlInputValue = "https://";
-    urlPromptVisible = true;
-    urlPromptCallback = (url) => {
-      if (url) {
-        try {
-          if (!url.startsWith("http://") && !url.startsWith("https://")) {
-            url = "https://" + url;
-          }
-          
-          editor?.createLink(url);
-          
-          selectionTooltipVisible = false;
-        } catch (e) {
-          toast.error("Failed to add link", { position: "bottom-end" });
-        }
-      }
-    };
-  }
-  
-  function submitUrlPrompt() {
-    if (urlPromptCallback) {
-      urlPromptCallback(urlInputValue);
-      urlPromptCallback = null;
-    }
-    urlPromptVisible = false;
-  }
-  
-  function cancelUrlPrompt() {
-    urlPromptCallback = null;
-    urlPromptVisible = false;
-  }
-
-  let selectionTooltipVisible = $state(false);
-  let selectionTooltipPosition = $state({ x: 0, y: 0 });
-  let formatCommands = $state([
-    { name: "Bold", icon: "tabler:bold", action: () => editor?.toggleStyles({bold: true}) },
-    { name: "Italic", icon: "tabler:italic", action: () => editor?.toggleStyles({italic: true}) },
-    { name: "Underline", icon: "tabler:underline", action: () => editor?.toggleStyles({underline: true}) },
-    { name: "Strike", icon: "tabler:strikethrough", action: () => editor?.toggleStyles({strike: true}) },
-    {name: "Link", icon: "tabler:link", action: () => addLinkToSelection() }
-    
-  ]);
-  let urlPromptVisible = $state(false);
-  let urlInputValue = $state("https://");
-  let urlPromptCallback = $state(null);
 </script>
 
 <header class="flex flex-none items-center justify-between border-b-1 pb-4">
@@ -838,57 +435,13 @@
             </div>
           {/if}
           <div class="relative">
-            <!-- TODO: get all users that has joined the server -->
             <ChatInput
               bind:content={messageInput}
               users={users()}
               context={contextItems}
               onEnter={sendMessage}
             />
-
-            <!--
-            {#if mayUploadImages}
-              <label
-                class="cursor-pointer text-white hover:text-gray-300 absolute right-3 top-1/2 -translate-y-1/2"
-              >
-                <input
-                  type="file"
-                  accept="image/*"
-                  multiple
-                  class="hidden"
-                  onchange={handleImageSelect}
-                />
-                <Icon
-                  icon="material-symbols:add-photo-alternate"
-                  class="text-2xl"
-                />
-              </label>
-            {/if}
-            -->
           </div>
-
-          <!-- Image preview 
-          {#if imageFiles?.length}
-            <div class="flex gap-2 flex-wrap">
-              {#each Array.from(imageFiles) as file}
-                <div class="relative mt-5">
-                  <img
-                    src={URL.createObjectURL(file)}
-                    alt={file.name}
-                    class="w-20 h-20 object-cover rounded"
-                  />
-                  <button
-                    type="button"
-                    class="absolute -top-2 -right-2 bg-red-500 rounded-full p-1"
-                    onclick={() => (imageFiles = null)}
-                  >
-                    <Icon icon="zondicons:close-solid" color="white" />
-                  </button>
-                </div>
-              {/each}
-            </div>
-          {/if}
-          -->
         </section>
       {/if}
 
@@ -1075,272 +628,12 @@
 
 {#snippet wikiTab()}
   {#if space}
-    <div class="flex flex-col gap-4 h-full overflow-y-auto p-2">
-      {#if !wikiContent || isEditingWiki}
-        <section class="wiki-editor-container">
-          {#if isEditingWiki}
-            <div class="mb-4 flex justify-between items-center">
-              <h3 class="text-xl font-bold text-white">{channel?.name} Wiki</h3>
-              <div class="flex gap-2">
-                <Button.Root
-                  onclick={() => setEditingWiki(false)}
-                  class="px-4 py-2 bg-gray-600 text-white rounded hover:bg-gray-500 transition-colors"
-                >
-                  Cancel
-                </Button.Root>
-                <Button.Root
-                  onclick={saveWikiContent}
-                  class="px-4 py-2 bg-violet-700 text-white rounded hover:bg-violet-600 transition-colors"
-                >
-                  Save
-                </Button.Root>
-              </div>
-            </div>
-            <div
-              class="wiki-editor bg-violet-900/20 rounded-lg border border-violet-500/30 p-4 h-auto {isAdmin ? 'admin-mode' : ''}"
-            >
-              <div bind:this={editorElement} class="min-h-[400px]"></div>
-
-              <!-- Slash Command Menu -->
-              {#if slashMenuVisible && isAdmin}
-                <div
-                  class="slash-menu bg-violet-900 border border-violet-700 rounded shadow-lg absolute z-50"
-                  style="left: {slashMenuPosition.x}px; top: {slashMenuPosition.y}px;"
-                >
-                  <ul class="py-1">
-                    {#each slashCommands as command}
-                      <li>
-                        <button
-                          class="flex items-center gap-2 w-full text-left px-4 py-2 hover:bg-violet-800 text-white"
-                          onclick={() => executeSlashCommand(command)}
-                        >
-                          <Icon icon={command.icon} class="text-xl" />
-                          <span>{command.name}</span>
-                        </button>
-                      </li>
-                    {/each}
-                  </ul>
-                </div>
-              {/if}
-
-              <!-- Selection Formatting Tooltip -->
-              {#if selectionTooltipVisible && isAdmin}
-                <div
-                  class="selection-tooltip bg-violet-900 border border-violet-700 rounded shadow-lg absolute z-50 flex"
-                  style="left: {selectionTooltipPosition.x}px; top: {selectionTooltipPosition.y}px;"
-                >
-                  {#each formatCommands as command}
-                    <button
-                      class="p-2 hover:bg-violet-800 text-white"
-                      title={command.name}
-                      onclick={() => {
-                        command.action();
-                        selectionTooltipVisible = false;
-                      }}
-                    >
-                      <Icon icon={command.icon} class="text-xl" />
-                    </button>
-                  {/each}
-                </div>
-              {/if}
-            </div>
-          {:else}
-            <div
-              class="flex justify-center items-center p-8 border border-dashed border-violet-500 rounded-lg"
-            >
-              <!-- Only show Create Wiki button to admins -->
-              {#if isAdmin}
-                <Button.Root
-                  onclick={() => setEditingWiki(true)}
-                  class="flex items-center gap-2 px-4 py-2 bg-violet-700 text-white rounded-lg hover:bg-violet-600 transition-colors"
-                >
-                  <Icon icon="tabler:edit" />
-                  Create Wiki Page
-                </Button.Root>
-              {:else}
-                <p class="text-gray-300">
-                  No wiki page exists yet. Only channel admins can create one.
-                </p>
-              {/if}
-            </div>
-          {/if}
-        </section>
-      {:else}
-        <section class="wiki-content">
-          <div class="flex justify-between items-center mb-4">
-            <h3 class="text-xl font-bold text-white">{channel?.name} Wiki</h3>
-            <!-- Only show Edit Wiki button to admins -->
-            {#if isAdmin}
-              <Button.Root
-                onclick={() => setEditingWiki(true)}
-                class="px-4 py-2 bg-violet-700 text-white rounded hover:bg-violet-600 transition-colors"
-              >
-                <Icon icon="tabler:edit" />
-                Edit Wiki
-              </Button.Root>
-            {/if}
-          </div>
-
-          <div class="wiki-rendered p-4 bg-violet-900/30 rounded-lg">
-            <div class="wiki-html text-white">
-              {@html wikiRenderedHtml}
-            </div>
-          </div>
-
-          
-        </section>
-      {/if}
-    </div>
+    {#if channel !== undefined && isAdmin !== undefined}
+      <WikiEditor
+        {space}
+        {channel}
+        {isAdmin}
+      />
+    {/if}
   {/if}
 {/snippet}
-
-{#if urlPromptVisible}
-  <div class="fixed inset-0 bg-black/50 flex items-center justify-center z-[100]">
-    <div class="bg-violet-900 border border-violet-700 rounded-lg shadow-lg p-6 max-w-md w-full">
-      <h3 class="text-lg font-bold text-white mb-4">Add Link</h3>
-      
-      <form 
-        onsubmit={submitUrlPrompt}
-        class="flex flex-col gap-4"
-      >
-        <input
-          type="text"
-          bind:value={urlInputValue}
-          placeholder="https://example.com"
-          class="w-full px-4 py-2 bg-violet-800 border border-violet-600 rounded text-white focus:outline-none focus:border-violet-500"
-          autofocus
-        />
-        
-        <div class="flex justify-end gap-3 mt-2">
-          <button
-            type="button"
-            class="px-4 py-2 bg-transparent border border-violet-500 text-white rounded hover:bg-violet-800 transition-colors"
-            onclick={cancelUrlPrompt}
-          >
-            Cancel
-          </button>
-          <button
-            type="submit"
-            class="px-4 py-2 bg-violet-700 text-white rounded hover:bg-violet-600 transition-colors"
-          >
-            Add Link
-          </button>
-        </div>
-      </form>
-    </div>
-  </div>
-{/if}
-
-<style>
-  :global(.bn-block) {
-    position: relative;
-    transition: padding-left 0.15s ease;
-    padding-left: 24px; 
-  }
-  
-  /* Only show the "+" button in the editor for admins */
-  :global(.wiki-editor-container .admin-mode .bn-block::before) {
-    content: '+';
-    position: absolute;
-    left: 6px;
-    top: 50%;
-    transform: translateY(-50%);
-    color: #a78bfa;
-    font-size: 16px;
-    font-weight: bold;
-    width: 18px;
-    height: 18px;
-    border-radius: 3px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    opacity: 0;
-    transition: opacity 0.15s ease, background-color 0.15s ease;
-    cursor: pointer;
-    z-index: 10;
-    background-color: transparent;
-  }
-  
-  :global(.wiki-editor-container .admin-mode .bn-block:hover::before) {
-    opacity: 1;
-  }
-  
-  :global(.wiki-editor-container .admin-mode .bn-block::before:hover) {
-    background-color: rgba(139, 92, 246, 0.2);
-  }
-  
-  
-  .bn-default-styles {
-    color: #fff;
-    padding-inline: 20px;
-  }
-
-  .slash-menu {
-    min-width: 200px;
-    max-width: 300px;
-    max-height: 300px;
-    overflow-y: auto;
-  }
-
-  .slash-menu ul {
-    max-height: 300px;
-    overflow-y: auto;
-  }
-
-  .wiki-editor {
-    overflow: visible;
-  }
-
-  :global(.wiki-rendered input[type="checkbox"]){
-    pointer-events: none;
-  }
-
-  :global(input[type="checkbox"]) {
-    pointer-events: auto; 
-    cursor: pointer;
-  }
-
-  :global(.bn-inline-content a){
-    color: #8b5cf6;
-    text-decoration: underline;
-    cursor: pointer;
-  }
-  
-  
-  :global(input[type="checkbox"]) {
-    accent-color: #8b5cf6;
-    width: 16px;
-    height: 16px;
-    margin-right: 8px;
-  }
-
-  .selection-tooltip {
-    padding: 4px;
-    border-radius: 6px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.3);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transform: translateX(-50%);
-  }
-  
-  .selection-tooltip button {
-    border-radius: 4px;
-    width: 36px;
-    height: 36px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transition: background-color 0.15s ease;
-  }
-  
-  /* Add a subtle animation for the tooltip */
-  .selection-tooltip {
-    animation: tooltipFadeIn 0.2s ease;
-  }
-  
-  @keyframes tooltipFadeIn {
-    from { opacity: 0; transform: translateY(5px) translateX(-50%); }
-    to { opacity: 1; transform: translateY(0) translateX(-50%); }
-  }
-</style>
